requirements.txt
flask 
google-generativeai 
python-dotenv
bs4
gunicorn